:sparkles: homeHeader UI 개선 (블러 처리)
