+++ Toss/Toss/Helper/DeviceInfo.swift
@@ -1,0 +1,12 @@
+//
+//  DeviceInfo.swift
+//  Toss
+//
+//  Created by Woody on 2022/07/05.
+//
+
+import Foundation
+
+struct DeviceInfo {
+    
+}
